struct Env made a little bit more sane.
There is no need for every Env instance to carry both a dictionary
and a parent-vector pair. It is either a local env or a global one.

TODO:
There are more to be done for struct Env. The API is misleading since
get_local/insert_local should be called only for a local env and
get_global/insert_global should be called only for a global env.